+++ app/src/main/java/com/downstairs/place/model/PlaceRepository.kt
@@ -4,7 +4,7 @@ import javax.inject.Inject
 
 class PlaceRepository @Inject constructor(private val placeDAO: PlaceDAO) {
 
-    suspend fun getPlace(placeId: Int) = placeDAO.findById(placeId)
+    suspend fun getPlace(placeId: Long) = placeDAO.findById(placeId)
 
     suspend fun insert() = placeDAO.insertPlace(Place(name = "EdFines", category = "Pub", description = "Nice"))
 }
